third_party/WebKit/LayoutTests/fast/forms/time-multiple-fields/time-multiple-fields-open-picker-key-bindings.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<script src="../resources/picker-common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<input type=time id=time value="2011-05-01">

<script>
description('F4 or Alt+Down should not open the calendar picker.');

var input = document.getElementById('time');
eventSender.keyDown('F4');
shouldBeNull('document.getElementById("mock-page-popup")');
eventSender.keyDown('F4', ['altKey']);
shouldBeNull('document.getElementById("mock-page-popup")');

finishJSTest();

</script>
</body>
</html>
